finish correction task
