+++ src/main/java/ru/itmentor/spring/boot_security/demo/controller/AdminController.java
@@ -2,13 +2,9 @@ package ru.itmentor.spring.boot_security.demo.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 import ru.itmentor.spring.boot_security.demo.model.User;
 import ru.itmentor.spring.boot_security.demo.service.UserService;
-import ru.itmentor.spring.boot_security.demo.service.UserServiceImpl;
 
 import java.util.List;
 
@@ -16,9 +12,9 @@ import java.util.List;
 @RequestMapping("/admin")
 public class AdminController {
 
-    private final UserServiceImpl userService;
+    private final UserService userService;
 
-    public AdminController(UserServiceImpl userService) {
+    public AdminController(UserService userService) {
 
         this.userService = userService;
     }
@@ -55,4 +51,9 @@ public class AdminController {
         userService.deleteUser(id);
         return "redirect:/admin/users";
     }
+@PostMapping("/updateRoles")
+    public String roles(@RequestParam("usId") int id,@RequestParam("roles") String roles){
+        userService.showUsersRoles(id,roles);
+        return "redirect:/admin/users";
+}
 }
