--- src/main/java/ru/itmentor/spring/boot_security/demo/security/RoleAuthority.java
@@ -1,20 +1,0 @@
-package ru.itmentor.spring.boot_security.demo.security;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public class RoleAuthority implements GrantedAuthority {
-    private String authority;
-
-    public RoleAuthority(String authority) {
-        this.authority = authority;
-    }
-
-    @Override
-    public String getAuthority() {
-        return authority;
-    }
-
-    public void setAuthority(String authority) {
-        this.authority = authority;
-    }
-}
