--- src/main/java/ru/itmentor/spring/boot_security/demo/security/UsersDetails.java
@@ -1,56 +1,0 @@
-package ru.itmentor.spring.boot_security.demo.security;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import ru.itmentor.spring.boot_security.demo.model.Role;
-import ru.itmentor.spring.boot_security.demo.model.User;
-
-import java.util.Collection;
-import java.util.stream.Collectors;
-
-public class UsersDetails implements UserDetails {
-
-    private final User user;
-    private final Collection<Role> roles;
-
-    public UsersDetails(User user, Collection<Role> roles) {
-        this.user = user;
-        this.roles = roles;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return roles.stream().map(r -> new SimpleGrantedAuthority(r.getAuthority())).collect(Collectors.toList());
-    }
-
-    @Override
-    public String getPassword() {
-        return this.user.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return this.user.getName();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
