+++ src/main/java/ru/itmentor/spring/boot_security/demo/service/UserDetailsServiceImpl.java
@@ -1,0 +1,27 @@
+package ru.itmentor.spring.boot_security.demo.service;
+
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Component;
+import ru.itmentor.spring.boot_security.demo.model.User;
+
+import java.util.stream.Collectors;
+@Component
+public class UserDetailsServiceImpl implements UserDetailsService {
+     private final UserService userService;
+
+    public UserDetailsServiceImpl(UserService userService) {
+        this.userService = userService;
+    }
+
+    @Override
+    @org.springframework.transaction.annotation.Transactional(readOnly = true)
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        User user = userService.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException(String.format("Пользователь не найден")));
+        return new org.springframework.security.core.userdetails.User(user.getName(),
+                user.getPassword(),
+                user.getRoles().stream().map(role -> new SimpleGrantedAuthority(role.getAuthority())).collect(Collectors.toList()));
+    }
+}
