+++ src/main/java/ru/itmentor/spring/boot_security/demo/service/UserService.java
@@ -1,70 +1,22 @@
 package ru.itmentor.spring.boot_security.demo.service;
 
-import org.hibernate.Hibernate;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import ru.itmentor.spring.boot_security.demo.model.Role;
 import ru.itmentor.spring.boot_security.demo.model.User;
-import ru.itmentor.spring.boot_security.demo.repository.RoleRepository;
-import ru.itmentor.spring.boot_security.demo.repository.UserRepository;
-import ru.itmentor.spring.boot_security.demo.security.UsersDetails;
 
-
-import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
 
-@Service
-public class UserService implements UserDetailsService {
-    private final UserRepository userRepository;
-    private final RoleRepository roleRepository;
-    private final PasswordEncoder passwordEncoder;
+public interface UserService {
 
-    public UserService(UserRepository userRepository, RoleRepository roleRepository, PasswordEncoder passwordEncoder) {
-        this.userRepository = userRepository;
-        this.roleRepository = roleRepository;
-        this.passwordEncoder = passwordEncoder;
-    }
+    public Optional<User> findByUsername(String username);
 
-    public Optional<User> findByUsername(String username) {
-        return userRepository.findByName(username);
-    }
+    public void saveUser(User user);
 
-    @Override
-    @org.springframework.transaction.annotation.Transactional(readOnly = true)
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = findByUsername(username).orElseThrow(() -> new UsernameNotFoundException(String.format("Пользователь не найден")));
-        return new org.springframework.security.core.userdetails.User(user.getName(),
-                user.getPassword(),
-                user.getRoles().stream().map(role -> new SimpleGrantedAuthority(role.getAuthority())).collect(Collectors.toList()));
-    }
+    public void deleteUser(Integer id);
 
-    public void saveUser(User user) {
-        Role roleUser = roleRepository.findByAuthority("ROLE_USER");
-        user.setRoles(List.of(roleUser));
-        user.setPassword(passwordEncoder.encode(user.getPassword()));
-        userRepository.save(user);
-    }
+    public List<User> showUsers();
 
-    public void deleteUser(Integer id) {
-        userRepository.deleteById(id);
-    }
+    public User getUser(int id);
 
-    public List<User> showUsers() {
-        return StreamSupport.stream(userRepository.findAll().spliterator(), false)
-                .collect(Collectors.toList());
-    }
-
-
-    public User getUser(int id) {
-        return userRepository.findById(id).orElse(null);
-    }
-
-
+    public void showUsersRoles(Integer id, String roles);
 }
