+++ src/main/resources/templates/allUser.html
@@ -25,6 +25,14 @@
                 <input type="hidden" name="usId" th:value="${usr.id}"/>
                 <button type="submit">Delete</button>
             </form>
+            <form th:action="@{/admin/updateRoles}" method="post">
+                <input type="hidden" name="usId" th:value="${usr.id}"/>
+                <select name="roles">
+                    <option th:value="'ROLE_USER'">USER</option>
+                    <option th:value="'ROLE_ADMIN'">ADMIN</option>
+                </select>
+                <button type="submit">Change Role</button>
+            </form>
         </td>
     </tr>
 </table>
